Make wireguard killswitch work even if excluded networks not specified
Right now, due to the `-n` checks, the wireguard killswitch does not work unless `KILLSWITCH_EXCLUDEDNETWORKS_IPV4` or `KILLSWITCH_EXCLUDEDNETWORKS_IPV6` are specified because the `REJECT` rule is not added, as far as I can tell.

I had to specify `KILLSWITCH_EXCLUDEDNETWORKS_IPV4: "10.42.0.0/16"` to make it work for my cluster. I think this PR should make that unecessary.